+++ tests/testthat/helpers/testers.R
@@ -2249,11 +2249,7 @@ vsel_tester <- function(
     soltrms <- setdiff(soltrms, soltrms_plus)
     soltrms <- c(soltrms, labels(terms(as.formula(paste(". ~", soltrms_plus)))))
   }
-  expect_true(
-    all(soltrms %in% split_formula(vs$refmodel$formula,
-                                   add_main_effects = FALSE)),
-    info = info_str
-  )
+  expect_true(all(soltrms %in% trms_universe_split), info = info_str)
 
   # ce
   if (with_cv && (valsearch_expected || cv_method_expected == "kfold")) {
@@ -2268,25 +2264,25 @@ vsel_tester <- function(
                 info = info_str)
   }
 
-  # pct_solution_terms_cv
+  # solution_terms_cv
   if (with_cv && isTRUE(vs$validate_search)) {
-    expect_true(is.matrix(vs$pct_solution_terms_cv), info = info_str)
-    expect_type(vs$pct_solution_terms_cv, "double")
-    expect_identical(dim(vs$pct_solution_terms_cv),
-                     c(solterms_len_expected, 1L + solterms_len_expected),
+    expect_true(is.matrix(vs$solution_terms_cv), info = info_str)
+    expect_type(vs$solution_terms_cv, "character")
+    if (identical(cv_method_expected, "kfold")) {
+      n_folds <- K_tst
+    } else {
+      n_folds <- nobsv
+    }
+    expect_identical(dim(vs$solution_terms_cv),
+                     c(n_folds, solterms_len_expected),
                      info = info_str)
-    expect_identical(colnames(vs$pct_solution_terms_cv),
-                     c("size", vs$solution_terms),
-                     info = info_str)
-    expect_identical(vs$pct_solution_terms_cv[, "size"],
-                     as.numeric(seq_len(solterms_len_expected)),
-                     info = info_str)
-    pct_nonsize_nms <- setdiff(colnames(vs$pct_solution_terms_cv), "size")
-    pct_solterms <- vs$pct_solution_terms_cv[, pct_nonsize_nms, drop = FALSE]
-    expect_false(anyNA(pct_solterms), info = info_str)
-    expect_true(all(pct_solterms >= 0 & pct_solterms <= 1), info = info_str)
+    # We need the addition of `NA_character_` because of subsampled LOO CV:
+    expect_true(
+      all(vs$solution_terms_cv %in% c(trms_universe_split, NA_character_)),
+      info = info_str
+    )
   } else {
-    expect_null(vs$pct_solution_terms_cv, info = info_str)
+    expect_null(vs$solution_terms_cv, info = info_str)
   }
 
   # nterms_max
@@ -2359,11 +2355,10 @@ smmry_tester <- function(smmry, vsel_expected, nterms_max_expected = NULL,
   expect_type(smmry, "list")
   expect_named(
     smmry,
-    c("formula", "family", "nobs_train", "pct_solution_terms_cv", "type_test",
-      "nobs_test", "method", "cv_method", "K", "validate_search",
-      "clust_used_search", "clust_used_eval", "nprjdraws_search",
-      "nprjdraws_eval", "search_included", "nterms", "selection",
-      "resp_oscale", "deltas"),
+    c("formula", "family", "nobs_train", "type_test", "nobs_test", "method",
+      "cv_method", "K", "validate_search", "clust_used_search",
+      "clust_used_eval", "nprjdraws_search", "nprjdraws_eval",
+      "search_included", "nterms", "selection", "resp_oscale", "deltas"),
     info = info_str
   )
 
@@ -2373,8 +2368,8 @@ smmry_tester <- function(smmry, vsel_expected, nterms_max_expected = NULL,
   expect_identical(smmry$family, vsel_expected$refmodel$family,
                    info = info_str)
   for (nm in c(
-    "nobs_train", "pct_solution_terms_cv", "type_test", "nobs_test", "method",
-    "cv_method", "K", "validate_search", "clust_used_search", "clust_used_eval",
+    "nobs_train", "type_test", "nobs_test", "method", "cv_method", "K",
+    "validate_search", "clust_used_search", "clust_used_eval",
     "nprjdraws_search", "nprjdraws_eval"
   )) {
     expect_identical(smmry[[nm]], vsel_expected[[nm]],
@@ -2584,5 +2579,62 @@ smmry_sel_tester <- function(
     }
   }
 
+  return(invisible(TRUE))
+}
+
+# A helper function for testing the structure of a `ranking` object as returned
+# by ranking().
+#
+# @param rk A `ranking` object as returned by ranking().
+# @param fulldata_expected The expected `fulldata` element of `rk`.
+# @param foldwise_expected The expected `foldwise` element of `rk`.
+# @param info_str A single character string giving information to be printed in
+#   case of failure.
+#
+# @return `TRUE` (invisible).
+ranking_tester <- function(rk, fulldata_expected, foldwise_expected, info_str) {
+  expect_s3_class(rk, "ranking", exact = TRUE)
+  expect_type(rk, "list")
+  expect_named(rk, c("fulldata", "foldwise"), info = info_str)
+  # Since we test elements `solution_terms` and `solution_terms_cv` already
+  # thoroughly in vsel_tester(), we can simply plug these into
+  # `fulldata_expected` and `foldwise_expected` and then test via identical():
+  expect_identical(rk[["fulldata"]], fulldata_expected, info = info_str)
+  expect_identical(rk[["foldwise"]], foldwise_expected, info = info_str)
+  return(invisible(TRUE))
+}
+
+# A helper function for testing the structure of a `props` object as returned by
+# props().
+#
+# @param pr A `props` object as returned by props().
+# @param cumulate_expected A single logical value indicating whether props() was
+#   run with `cumulate = TRUE` (`TRUE`) or not (`FALSE`).
+# @param nterms_max_expected A single numeric value as supplied to
+#   props.ranking()'s argument `nterms_max`.
+# @param info_str A single character string giving information to be printed in
+#   case of failure.
+#
+# @return `TRUE` (invisible).
+props_tester <- function(pr, cumulate_expected = FALSE, nterms_max_expected,
+                         cnms_expected, info_str) {
+  classes_expected <- "props"
+  if (cumulate_expected) {
+    classes_expected <- c("cumulprops", classes_expected)
+  }
+  expect_s3_class(pr, classes_expected, exact = TRUE)
+  expect_equal(dim(pr), c(nterms_max_expected, nterms_max_expected),
+               info = info_str)
+  expect_true(is.numeric(pr), info = info_str)
+  rnms_expected <- as.character(seq_len(nterms_max_expected))
+  if (cumulate_expected) {
+    rnms_expected <- paste0("<=", rnms_expected)
+  }
+  if (length(cnms_expected) > nterms_max_expected) {
+    cnms_expected <- head(cnms_expected, nterms_max_expected)
+  }
+  expect_identical(dimnames(pr),
+                   list("size" = rnms_expected, "predictor" = cnms_expected),
+                   info = info_str)
   return(invisible(TRUE))
 }